LayoutTests/http/tests/paymentrequest/updateWith-log-error-if-not-synchronous.https-expected.txt

CONSOLE MESSAGE: updateWith() should be called synchronously when handling "shippingaddresschange".

PASS Not calling `updateWith` inside `shippingaddresschange` should log an error.
